Add burn and rewite burn test from ERC777 to ERC20
