+++ contracts/KTYtoken.sol
@@ -1,14 +1,18 @@
 pragma solidity ^0.5.0;
 
-import './openzeppelin_v2_4_0/token/ERC20/ERC20.sol';
-import './openzeppelin_v2_4_0/token/ERC20/ERC20Detailed.sol';
-import './openzeppelin_v2_4_0/ownership/Ownable.sol';
+import './openzeppelin_v2_5_0/token/ERC20/ERC20.sol';
+import './openzeppelin_v2_5_0/token/ERC20/ERC20Detailed.sol';
+import './openzeppelin_v2_5_0/token/ERC20/ERC20Burnable.sol';
+import './openzeppelin_v2_5_0/ownership/Ownable.sol';
+import './openzeppelin_v2_5_0/math/SafeMath.sol';
 
-contract KTYtoken is Ownable, ERC20, ERC20Detailed {
+contract KTYtoken is Ownable, ERC20, ERC20Detailed, ERC20Burnable {
+    using SafeMath for uint256;
+
     string constant NAME    = 'Kittiefight';
     string constant SYMBOL  = 'KTY';
-    string constant DECIMALS  = 18;
-    uint256 constant MAX_TOTAL_SUPPLY = 100_000_000 * 10**DECIMALS;
+    uint8 constant DECIMALS  = 18;
+    uint256 constant MAX_TOTAL_SUPPLY = 100_000_000 * 10**uint256(DECIMALS);
 
     constructor() ERC20Detailed(NAME, SYMBOL, DECIMALS) public {
     }
@@ -22,7 +26,7 @@ contract KTYtoken is Ownable, ERC20, ERC20Detailed {
      */
     function mint(address account, uint256 amount) public onlyOwner returns (bool) {
         require(totalSupply().add(amount) <= MAX_TOTAL_SUPPLY, "KTYtoken: cap exceeded");
-        _mint(_msgSender(), account, amount, '', '');
+        _mint(account, amount);
         return true;
     }
 }
